pyproject.toml
[project]
name = "filed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-user-pt>=0.6.21",
]
